+++ src/application/hooks/api/index.ts
@@ -1,5 +1,4 @@
 export * from "./account";
 export * from "./collection";
-export * from "./meme";
 export * from "./search";
 export * from "./tags";
