--- src/application/hooks/api/meme/queryKey.ts
@@ -1,11 +1,0 @@
-import { CORE_QUERY_KEY } from "@/api/core";
-
-export const QUERY_KEYS = {
-  getMemeDetailById: (id: string) => ["getMemeDetailById", id],
-  getMemesBySort: (sort: string) => [CORE_QUERY_KEY.infiniteMemeList, "getMemesBySort", sort],
-  getMemesByCollectionId: (collectionId: number | null) => [
-    CORE_QUERY_KEY.infiniteMemeList,
-    "getMemesByCollectionId",
-    collectionId,
-  ],
-} as const;
